==> notebooks/01_exploration_spark.ipynb <==
#### Etape 1.1 : Exploration initiale
- Charger les donnees de consommation avec PySpark
- Analyser le schema infere et identifier les problemes de typage
- Calculer les statistiques descriptives par type d'energie
- Identifier les batiments avec le plus de mesures
- Produire un rapport d'audit de qualite des donnees

**Livrables** :
- Notebook `01_exploration_spark.ipynb`
- Rapport d'audit (format markdown ou DataFrame)

from pyspark.sql import SparkSession
import pyspark.sql.functions as F

spark = SparkSession.builder.master("spark://spark-master:7077").appName("01_exploration").getOrCreate()

# Charger les donnees de consommation avec PySpark
df = spark.read \
          .option("header", "true") \
          .option("inferSchema", "true") \
          .option("sep", ",") \
          .csv("../data_ecf/consommations_raw.csv")

# Analyser le schema
df.printSchema()
df.show()

# Calculer les statistiques descriptives par type d'energie
df.groupBy("type_energie").agg(
    F.count("*").alias("count"),
    F.round(F.sum("consommation"),2).alias("total_consommation_kWh"),
    F.round(F.avg("consommation"),2).alias("avg_consommation_kWh")
).show()

# Identifier les batiments avec le plus de mesures
df.groupBy("batiment_id").count().orderBy(F.desc("count")).show()

# Produire un rapport d'audit de qualite des donnees
print("Nombre de lignes total : ", df.count())
print("Nombre de lignes avec valeurs manquantes : ", df.filter(F.col("consommation").isNull()).count())
print("Nombre de lignes avec valeurs nuls : ", df.filter(F.col("consommation") == 0).count())

spark.stop()
